requirements.txt
aiofiles==23.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
Jinja2==3.1.4
openai==1.35.10
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.30.1
websockets==12.0
streamlit
watchdog